XAP-13017 Fix ConcurrentModificationException
ConcurrentModificationException might be thrown when
iterating through the newNotifies array.

The handleThrowable is called by a callback from
LRMI async future result, which accesses the
newNotifies array - it should always be accessed
under a writeLock.

The onResult is not called by the same LRMI thread,
thus there is no risk in reentrant deadlock.

Reviewer: Niv